refactor: PSN-12 Refactoring, update deps and update to go 1.21
