--- .circleci/config.yml.notinuse
@@ -1,40 +1,0 @@
-version: 2
-
-jobs:
-  test:
-    docker:
-      - image: cimg/go:1.17
-    steps:
-      - checkout
-      - restore_cache:
-          keys:
-            - go-mod-v1-{{ checksum "go.sum" }}
-      - run:
-          name: Install Dependencies
-          command: go get ./...
-      - save_cache:
-          key: go-mod-v1-{{ checksum "go.sum" }}
-          paths:
-            - "/go/pkg/mod"
-      - run:
-          name: Run tests
-          command: go test -v ./...
-  deploy:
-    docker:
-      - image: arvindr226/alpine-ssh
-    steps:
-      - checkout
-      - run: ssh -oStrictHostKeyChecking=no -v $USER@$IP "./deploy-puasa-sunnah.sh"
-
-workflows:
-  version: 2
-  workflow:
-    jobs:
-    - test
-    - deploy:
-        filters:
-          branches:
-            only:
-              - main
-        requires:
-          - test
