+++ .golangci.yml
@@ -1,0 +1,3 @@
+linters:
+  enable:
+    - gocritic
