+++ src/service/source.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"github.com/granitebps/puasa-sunnah-api/src/repository"
-	"github.com/granitebps/puasa-sunnah-api/types"
+	"github.com/granitebps/puasa-sunnah-api/src/types"
 )
 
 type SourceService struct {
